Revision on Python solutions about print only answer finished! #2
Now we can to type: python stats.py -s Python --build

If any do that, we'd get the time of each execution in
python solutions and your answer!

But I doesn't have sure if all the solutions are correct...

The next goal with Python will be generating hashes for each solution
to check if are correct.