Make Common Lisp solution for Problem045 portable to SBCL
We have a problem with floating point error and different implementations
of sqrt routine between SBCL and CLISP.

On SBCL sqrt never returns integers, on CLISP yes.

SBCL:
(sqrt 4) => 2.0
CLISP:
(sqrt 4) => 2

The old solution is a problem because it expects that somehow a integer
returns, but on SBCL this is a infinite loop (sqrt of SBCL never returns
an integer).

To fix that I rewrite the main solution function and use coerce for
high precision float number: 'long-float.

This fixed and the solution works on both compilers.