Add analytical solution for Problem 102 written in Haskell
Probably is my first solution in Haskell here.

Time: 0.06s!

Yes. Just this.

This solution was made without brute-force,
for first solved just using pen & paper.

The central idea is using inequalities for each segment of the triangle
based on linear functions created with the points. For each triangle
I generate three inequalities which will be applied to arbitrary points
later to check if that points is inside of the triangle.

	^ (y)
	|            B
	|           /\
	|     f(x) /  \ g(x)
	|         /  c \
	|        /______\
	|       A        C
	|          h(x)
	|----------------------> (x)

Intuition:
Given a ABC triangle and its sides AB, BC, AC,
find the coincident functions f(x), g(x), h(x) for each segment.
Based on that functions, write a inequality with the opposite vertex
of the segment to delimit the plane; applying all the inequalities
from a given point we can know if that points is inside or not of
the triangle.

Works pretty well.