Trying again fix the C build on travis: fixes #13
The problem is:

Maybe the version of gcc on travis need to have the link libraries
on the end of the command. On the earlier version of the stats.py
we have only the way to pass params as prefixed before the
`-o compiled.out` as defined on Builder class using the
BUILD_MACHINE[language]['cmdline'].

Now, optionally, the Execute builder can have the ext_params.