[.travis.yml] force install clojure1.3
For some weird reason now the Ubuntu LTS 12.04 on Travis
doesn't find anymore triggers to install the clojure candidate. Weird.

I'm following this guide:
https://www.thelinuxfaq.com/ubuntu/ubuntu-12-04-lts-precise-pangolin/clojure1.3

Would be better if I have a Ubuntu 12.04 to test or if I can
plug my terminal on travis to debug.

This seem possible with this:
https://docs.travis-ci.com/user/running-build-in-debug-mode/

But for public repositories is a little annoying to setup.