[.travis.yml] Export TRAVIS_PULL_REQUEST variable to docker-compose
Docker is a sandbox/container service which normally is executed
on a isolated environment, so we need explicitly pass that type
of environment variables to docker-compose or what else docker wrapper.

BTW, we can simplify this painful process calling internally
docker-compose of the 'test' file instead. But this will break the
possible usage of someone use the 'test' script without docker and
this is bad. No one have a proper installation of docker because
docker installation is pain. Only people which uses this have docker.

So I'd like to avoid that.